Support panic backtraces, add log buffer
When a test fails there is no way to use RUST_BACKTRACE=1 to show a
backtrace. This patch fixes the issue.

Additionally, it adds an explicit log buffer that can be used to work
around #9. That's a feature that the std harness doesn't have, but I
think it's appropriate to add since it's a workaround for something
libtest-mimic _doesn't_ but should have.